requirements.txt
dbt-adapters==1.23.0
dbt-common==1.38.0
dbt-core==1.11.10
dbt-extractor==0.6.0
dbt-protos==1.0.498
dbt-semantic-interfaces==0.9.0
dbt-snowflake==1.11.5
pandas==3.0.3
plotly==6.7.0
python-dotenv==1.2.2
snowflake-connector-python==4.5.0
streamlit==1.57.0
